+++ src/azure-cli-core/HISTORY.rst
@@ -3,7 +3,19 @@
 Release History
 ===============
 
+2.0.39
+++++++
+* Minor fixes
+
+2.0.38
+++++++
+* Add global support for `--subscription` to most commands.
+
 2.0.37
+++++++
+* Minor fixes
+
+2.0.36
 ++++++
 * Minor fixes